requirements.txt
sstreamlit
pandas 
numpy
reportlab
streamlit-aggrid